README.md
hat are the benefits of neural networks for motor tariffing? To answer this question, claim frequencies are modeled and predicted for a large French motor third party liability insurance portfolio. In a first classical approach generalized linear models (GLM) as well as their mixed model cousin (GLMM) are used. Then this approach is extended to deep artificial neural networks and the novel combined actuarial neural net (CANN) is used in the implementation of Schelldorfer and Wüthrich (2019). Subsequently decision-tree-based model ensembles (eXtreme Gradient Boosting, "XGBoost") are applied and the models examined. In addition, questions of tariff structures and model stability are regarded and cross-validation is carried out. It is shown that deep neural networks as well as decision tree based model ensembles can be used at least for the improvement of classical models. Furthermore, XGBoost models prove to be the superior forecasting models, taking into account the tariff system.